add app package
